requirements.txt
langchain
langchain-groq
yfinance
python-dotenv